Task 4

+++ alayatodo/views.py
@@ -49,7 +49,7 @@ def logout():
 def todo(id):
     if not session.get('logged_in'):
         return redirect('/login')
-    cur = g.db.execute("SELECT * FROM todos WHERE id ='%s'" % id)
+    cur = g.db.execute("SELECT * FROM todos WHERE id = '%s' AND user_id = '%s'" % (id, session.get('user')['id']))
     todo = cur.fetchone()
     return render_template('todo.html', todo=todo)
 
@@ -59,9 +59,9 @@ def todo(id):
 def todos():
     if not session.get('logged_in'):
         return redirect('/login')
-    cur = g.db.execute("SELECT * FROM todos")
+    cur = g.db.execute("SELECT * FROM todos WHERE user_id = '%s'" % session.get('user')['id'])
     todos = cur.fetchall()
-    return render_template('todos.html', todos=todos)
+    return render_template('todos.html', todos=todos, get_flashed_messages=get_flashed_messages)
 
 
 @app.route('/todo', methods=['POST'])
@@ -77,23 +77,25 @@ def todos_POST():
         % (session['user']['id'], description)
     )
     g.db.commit()
-    return redirect('/todo')
-
-
-@app.route('/todo/<id>', methods=['DELETE'])
-def todo_DELETE(id):
-    if not session.get('logged_in'):
-        return redirect('/login')
-    g.db.execute("DELETE FROM todos WHERE id ='%s'" % id)
-    g.db.commit()
+    session['alert'] = ['A todo has been successfully created!']
     return redirect('/todo')
 
 
 @app.route('/todo/<id>', methods=['POST'])
+def todo_DELETE(id):
+    if not session.get('logged_in'):
+        return redirect('/login')
+    g.db.execute("DELETE FROM todos WHERE id ='%s' AND user_id = '%s'" % (id, session.get('user')['id']))
+    g.db.commit()
+    session['alert'] = ['A todo has been successfully deleted!']
+    return redirect('/todo')
+
+
+@app.route('/todo/<id>/complete', methods=['POST'])
 def todo_POST(id):
     if not session.get('logged_in'):
         return redirect('/login')
-    g.db.execute("UPDATE todos SET completed=1 WHERE id ='%s'" % id)
+    g.db.execute("UPDATE todos SET completed=1 WHERE id ='%s' AND user_id = '%s'" % (id, session.get('user')['id']))
     g.db.commit()
     return redirect('/todo')
 
@@ -102,5 +104,11 @@ def todo_POST(id):
 def todo_json(id):
     if not session.get('logged_in'):
         return redirect('/login')
-    cur = g.db.execute("SELECT * FROM todos WHERE id ='%s'" % id)
+    cur = g.db.execute("SELECT * FROM todos WHERE id ='%s' AND user_id = '%s'" % (id, session.get('user')['id']))
     return jsonify(dict(zip([column[0] for column in cur.description], cur.fetchone())))
+
+
+def get_flashed_messages():
+    alert = session['alert'] if 'alert' in session else []
+    session.pop('alert', None)
+    return alert
